% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{loadFile}
\alias{loadFile}
\title{Load File into object}
\usage{
loadFile(x, y)
}
\arguments{
\item{x}{An object}

\item{y}{A file name}
}
\value{
The object containing the data read in from the file
}
\description{
Load File into object
}
\examples{
ddt = loadFile(ddt, "DDT.csv")
}
