% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Bootstrap}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of iterations}

\item{x}{data set}

\item{fun}{function}

\item{alpha}{alpha}

\item{cx}{cx for hist}
}
\value{
List of the bootstrap
}
\description{
Bootstrap
}
\examples{
myboot(iter = 10000,x = sam,fun="mean",alpha=0.05,cx=1.5)
}
