% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poisson distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Size}

\item{iter}{Number of iterations}

\item{lambda}{Lambda}
}
\value{
Histogram of the sample
}
\description{
Poisson distribution
}
\examples{
w = mycltp(1,10000,10)
}
