% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{MyMaxLik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function to use}

\item{x}{A vector}

\item{param}{params for plot}
}
\value{
The list and plot of the max likelyhood
}
\description{
MyMaxLik
}
\examples{
 mymaxlik(logbin, x, seq(0,1,1000))
}
