% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymult.R
\name{mymult}
\alias{mymult}
\title{mymult}
\usage{
mymult(iter = 100, n = 10, p = c(1, 1, 1, 1)/4)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Size}

\item{p}{A file name}
}
\value{
Plot of a multinomial
}
\description{
mymult
}
\examples{
mymult(iter=1000,n=10,p=c(1,2,3,4,2)/12)
}
