% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{N Curve Plot and Area}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Average}

\item{sigma}{Standard Deviation}

\item{a}{Upper Bound}
}
\value{
Graph of curve and area of the shaded area between curve and x axis
}
\description{
N Curve Plot and Area
}
\examples{
myncurve(mu=10,sigma=5, a=6)
}
